Add condensed EEPROM blocks
The prop c generated is a bit ugly, but I have two requests into Simple Libraries that will bring it down to line or two per block.  I'm happy to change this code again one those are in place.  These blocks make sure the user doesn't overwrite program or data already stored in the EEPROM, and it makes the string handling a lot more intuitive by taking advantage of the null terminators that strings are supposed to have.